Interned frequently-used strings in generic code
